+++ tests/helpers/fake_log_data.h
@@ -91,7 +91,7 @@ protected:
             static_cast<LineLength::UnderlyingType>( doGetLineString( line ).size() ) );
     }
     void doSetDisplayEncoding( const char* ) override {}
-    QTextCodec* doGetDisplayEncoding() const override
+    const TextEncoding* doGetDisplayEncoding() const override
     {
         return nullptr;
     }